+++ test/v2/createEntries.test.js
@@ -71,73 +71,3 @@ describe('createTickEntries', () => {
     });
   });
 });
-
-describe('listTickEntries', () => {
-  const params = {
-    startDate: '2021-11-08',
-    endDate: '2021-11-09',
-    userId: 'userId',
-  };
-
-  const baseURL = 'TICK_BASE_URL_START/123456/api/v2/entries.json';
-  const defaultHeaders = {
-    Authorization: 'Token token=ar425598573462y24ec1ceee728981663',
-    'User-Agent': 'tickspot.js (user@kommit.co)',
-  };
-
-  describe('when API call is successful', () => {
-    const responseData = [{ id: '234' }, { id: '235' }];
-    beforeEach(() => {
-      axios.get.mockClear();
-      axios.get.mockResolvedValueOnce({ data: responseData });
-    });
-
-    it('should return the list of tick entries', async () => {
-      const response = await client.entries.list(params);
-
-      expect(axios.get).toHaveBeenCalledTimes(1);
-      expect(axios.get).toHaveBeenCalledWith(baseURL,
-        {
-          headers: defaultHeaders,
-          params: { end_date: '2021-11-09', start_date: '2021-11-08', user_id: 'userId' },
-        });
-      expect(response).toEqual([{ id: '234' }, { id: '235' }]);
-    });
-  });
-
-  describe('when API call returns an error', () => {
-    const responseError = 'test';
-    beforeEach(() => {
-      axios.get.mockClear();
-      axios.get.mockRejectedValueOnce(responseError);
-    });
-
-    it('an error should be thrown when making the call', async () => {
-      await expect(client.entries.list(params)).rejects.toThrow(responseError);
-
-      expect(axios.get).toHaveBeenCalledTimes(1);
-      expect(axios.get).toHaveBeenCalledWith(baseURL,
-        {
-          headers: defaultHeaders,
-          params: { end_date: '2021-11-09', start_date: '2021-11-08', user_id: 'userId' },
-        });
-    });
-  });
-
-  describe('when the list method is not sent any parameter', () => {
-    const responseData = [{ id: '234' }, { id: '235' }];
-    beforeEach(() => {
-      axios.get.mockClear();
-      axios.get.mockResolvedValueOnce({ data: responseData });
-    });
-
-    it('should throw an error specifying which parameter is missing', async () => {
-      await expect(client.entries.list({})).rejects.toThrow('startDate field is missing');
-
-      await expect(client.entries.list({ ...params, endDate: null }))
-        .rejects.toThrow('endDate field is missing');
-
-      expect(axios.get).not.toHaveBeenCalled();
-    });
-  });
-});
