+++ test/v2/fixture/entries/listEntriesResponseData.js
@@ -1,0 +1,26 @@
+const dataEntriesSuccessful = {
+  data: [{
+    id: 123456,
+    date: '2021-11-08',
+    hours: 0.2133,
+    notes: '[OpenSource] Software Development',
+    task_id: 14541266,
+    user_id: 337683,
+    url: 'https://secure.tickspot.com/114217/api/v2/entries/123456.json',
+    created_at: '2021-11-08T11:16:37.000-05:00',
+    updated_at: '2021-11-08T11:16:37.000-05:00',
+  },
+  {
+    id: 123457,
+    date: '2021-11-09',
+    hours: 0.3642,
+    notes: '[OpenSource] Software Development',
+    task_id: 14541266,
+    user_id: 337683,
+    url: 'https://secure.tickspot.com/114217/api/v2/entries/123457.json',
+    created_at: '2021-11-09T11:16:37.000-05:00',
+    updated_at: '2021-11-09T11:16:37.000-05:00',
+  }],
+};
+
+export default dataEntriesSuccessful;
